requirements.txt
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==2.3.7
psycopg2-binary==2.9.9
requests==2.31.0
